std: separate out bench file
